Remove interleaved synapse writing
The interleaved (aka parallel) synapse writing was beneficial
in the busy-wait-mode.

The addition of the wait-by-dummy (aka (dummy-wait or min-delay)
mode, makes it write-time wise unneeded.

In addition, to fully profit from interleaved programming with
with the wait-by-dummy-mode, the minimum number of dummy packets
needed would have to be adjusted as the current implementation
always uses the number of the not-interleaved scenario.

Therefore, the interleaved mode is hereby removed.

Change-Id: I1255f4ac9e2bac71be1ca414efc5880ad3a7fcfe